requirements.txt
google-generativeai
speechrecognition
pyaudio
python-dotenv
